stm_fdma: Fix callers of FDMA dma_compile_list() to use GPF_ATOMIC in irq
This fixes a problem introduced by the
  linux-2.6.17.14_stm23ear_0100-fdma-rework.patch
patch.

The fdma-rework patch removed the parameter to declare_dma_parms() which
indicated whether future operations on those parameters should be
assumed to come from interrupt context. This was moved to a new parameter
to dma_compile_list(), which only applied to that call.

However, this was forgotten about when updating some of the drivers
which use this API, and functions which are called from interrupt context
were incorrectly modified to call dma_compile_list() with GFP_KERNEL
when it should have been GFP_ATOMIC. This patch fixes this problem,
and also a missing spinlock initialisation.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>