stm_copro: Fix coprocessor driver watchdog timer reset problem
This fixes a problem seen on the 710x, where the system crashes
shortly after a reboot (using either the reboot command or watchdog), but
is fine when hitting the reset button.

The problem with the reset is that after the st231 firmware is loaded it leaves
the reset chain such that a watchdog reset will not cause the on-chip
peripherals to be reset. This patch fixes the reset chaining problem, and
allows a watchdog reset to work correctly.

This resolves bugzilla 1614

Signed-off-by: Peter Bennett <peter.bennett@st.com>