sh: Changes the way that gUSA regions are handled for SH
Originally the kernel checked to see if it had interrupted a gUSA
sequence only in the case of a returning to user code via the scheduler,
or in the case of a signal handler. This worked fine, but a known
limitation was that it wouldn't work with preemption enabled.

In an attempt to overcome this an extra test was added: when __switch_to()
was called it also checked for an interrupted gUSA sequence. Unfortunately
this has two bugs:

 - a minor problem was that the kernel updated the user stack pointer
   to the pre-gUSA value, but only updated the user PC to rollback point
   #2 (after the gUSA code modifies the stack pointer). Thus if the
   user code started to re-execute the sequence after rollback and was then
   interrupted a second time, it would not be detected as being in a
   gUSA sequence, and so would not be rolled back when it should.

 - more problematically the code used user_regs() to detect whether the
   task was a user task. However in the case of kernel threads, there is
   no pt_regs structure at the top of the kernel stack, and so the
   code could erroneously think that the interrupted code was in the middle
   of a gUSA sequence, rollback the registers, and so corrupt the
   kernel thread's registers.

While both of these problems could be solved by simply updating the code,
this patch takes a slightly different approach.

On every entry to the kernel (interrupt, exception or trap) it checks
to see if the interrupted code was in the middle of a gUSA sequence, and
if so rolls it back. This has a number of advantages:

 - it means that the check only has to be performed in one place, and the
   rest of the kernel is unaware of gUSA handling. This reduces the
   chance of errors, as there is no chance of missing a path through the
   kernel code where a check would be required.

 - at this point all the state the code needs to check and update is
   held in registers rather than memory, making the check less expensive.

 - it gives the potential to use the rollback technique in kernel space
   as well as user space.

It does have a number of disadvantages:

 - the check is performed more frequently, so the effect on kernel performance
   should probably be investigated.

 - the code always readjusts the stack pointer and updates the PC to
   rollback point #1, which means on restarting a rolled back sequence
   one more instruction has to be executed.

If the principle is accepted, then it would probably be worth investigating
an optimisation which overlaps the gUSA checks with determining whether the
interrupted thread was in user or kernel space. This might make the code
less maintainable, but is probably worthwhile.
   
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>