cdrom: Fix ATAPI packet commands using DMA buffers on the stack
An intermittent problem has been seen using CD-ROM and DVD drives attached
using SATA and USB interfaces. Most frequently this is seen while mounting
the filesystem, but it can occur when reading from the disk. The usual
result is a kernel panic, where the stack dump shows file system related
operations, but the PC is garbage.

This problem is caused by a bug in the code which receives data using the
ATAPI PACKET command. This data may be transferred from the device using
DMA, and so has to be in a block of memory where the device driver can
safely perform cache invalidation. The current code simply passes in the
address of a buffer on the stack, so if the device driver performs a cache
invalidate on this block, adjacent words may well also be invalidated.
Frequently this includes the function's return address, which is corrupted,
and so the function returns to a garbage address.

This patch modifies init_cdrom_command() to allocate the buffer (if size
is non-zero). It also uses the gcc "cleanup" attribute on the packet_command
structure to ensure that the buffer is freed when the function exits.

SATA and USB CD-ROM devices now appear to work correctly. The patch also
touches some IDE CD-ROM code (which has had a small amount of testing), and
packet writing (which has had no testing at all).

This resolves bugzilla 1478.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>