sh: Add sys_cacheflush() call for SH CPUs.
Adds a system call to allow user code to flush code from the cache.
You can use instructions for the data side, but the iside can 
only be done by a flush ROM which really only works with a direct 
mapped cache. The later SH4's have 2 way Iside, so this call allows 
a portable way to flush the cache.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>