sh_stm: Make 710x HDMI SSC port I2C only, it cannot be used as SPI
This patch changes the SSC platform configuration on all ST 710x boards such
that the SSC port containing the HDMI detect line is I2C only. 

This is because on all of the affected boards the HDMI detect is wired into
the same pin as would be used by an SPI data line.  As the SSC port is only
brought out to a test header, it is not available for normal use and hence
only the two-wire I2C bus is available.

The patch also rewrites the SSC configuration setup to use the symbolic 
constants, rather than just a hex number, so the configuration is human 
readable.

Signed-off-by: Stephen Gallimore <stephen.gallimore@st.com>