sh_stm: Various tidy ups for STmicro mb519 7200 board
Fix some problems with the I2C/SPI bus defintions on the mb519:
 - SSC0 is connected to the HDMI, so effectivly dedicated to I2C
 - SSC1 and SSC3 connect through to the daughter board, and so can
   be either I2C or SPI, so allow them to be either.

Also print the board and EPLD revision numbers in the correct format.

Add support for the Flash device.

Add support for the heartbeat LED on the mb520.

Turn off the bypass of ClockGenB to make USB 1.1 work.

Finally disable the second Ethernet device by default. This shouldn't
be necessary, but because there is missing pull up resistor on the MDIO
line, we can't tell the difference between when a PHY is present and when
it isn't. Unfortuntaly the MAC crashes if you try to reset it when it
isn't receiving a clock from the PHY, and so the net result is you
should only enable it when a PHY is present, and currently that is rare.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>