i2c-stm: Update bus numbering in ST40 PIO I2C driver
This patch makes the I2C PIO driver consistent with the SSC I2C driver
in that on the MB442 CoCo ref board, only two I2C busses are usable.

The third bus is permanently connected to a non I2C device and is 
disabled by this patch.

This mirrors the patch linux-sh4-2.6.11_stm20-31.i2c_7100ref_fix.patch,
which fixes the SSC driver.

Signed-off-by: Carl Shaw <carl.shaw@st.com>