stm_pwm: Add driver for PWM hardware present in many STMicro chips.
This is a simple driver for the PWM (Pulse Width Modulation) hardware
present on many ST chips. Although the PWM hardware can be used for
many things, this driver simply allows the mark/space ratio to be set,
which is sufficient for many purposes, including controlling the speed
of a fan.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>