alsa_stm: Fix idempotency problems in snd_pcm_ops::prepare
The recent FDMA rework patch introduced a bug into the prepare
operation for the STb7100 PCM playback and capture devices.
Specifically it now requests a DMA req line every time the soundcard
is prepared (and on the second attempt to request the line it is,
unsuprisingly, busy).

This patch corrects the problem.

Signed-off-by: Daniel Thompson <daniel.thompson@st.com>