sh: Make UBC register accesses 8 bit instead of 32 bit
On at least some parts (including 7750, 7751 and ST40-20x) the UBC.BAMRA
register needs to be addressed as an 8 bit register.

Its not clear why this hasn't been a problem until now, but in a very
limited set of parts it appears the value in the register ends up corrupt,
causing strange breakpoint behaviour which usually results in a chip
lock up.

Signed-off-by: Chris Smith <chris.smith@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>