sh_stm: Update USB power pin setting
Until this patch was added, we relied on the stb710x ref board to power
up with the USB power distribution chip enabled. It was not actively driven
by the 710x at all. However rev C of the board changed the power on state,
so we now have to drive it, but because the enables signal polarity doesn't
match that generated by the 710x USB hardware, we have to drive it as a PIO
pin.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>