smsc911x: Enable use of linked paced DMA transfers
This patch adds support to the SMSC 911x driver to use linked paced
transfers rather than free-running DMA to receive data.

This is mainly used as a work-around for broken boards, which do not
have the FIFOSEL pin wired up. On these we have no choose but to read
from the main register bank, where the receive FIFO is mapped multiple
times. We can do this using free-running 2D transfers, but these are
comparatively slow. However paced transfers, with a large read and
non-incrementing address have the same effect, and comparable performance
to FIFOSEL operation.

To do this the driver needs two pacing signals which are
permanently active, by default this is SSC 1 and 2 Tx empty, but
this may need to be changed if either SSC is in use.

A special mode is available in some FDMA firmware versions, where channel
11 has additional buffering. In theory this should give improved
performance, and so the FDMA driver has been modified to reserve this
channel for Ethernet use.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>