sh_stm: Improve platformisation of STMicro 7100 USB wrapper
The current USB support had all the device register addresses within
the USB code, which makes it dificult to port to different platforms. It
also assumes that there is only one OHCI/EHCI controller combo, which
breaks when there is more than one.

To address this, move all the address management into the BSP.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>