video: Check return values of framebuffer set video mode.
I have discovered that the generic framebuffer code which actually sets 
a video mode ignores the return code from the driver specific call. This
means that if setting the video mode fails the generic code will blindly
continue and assume the new video mode is in effect when it isn't. It
appears that somewhere between 2.4 and the current 2.6 the FB driver
API changed, previously the equivalent 2.4 routine had no return value.
All other drivers implement the new interface and at least two (Matrox and
SiS) can return errors. It looks like an oversight that the generic code
hasn't been changed to reflect this, the attached bk diff is my proposed
solution.

Signed-off-by: Stephen Gallimore <stephen.gallimore@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>