sh: Implement atomic operations with rollback technique for SuperH processors.
This patch implements most of the kernel atomic operations using sofware
rollback, rather than simply disabling interrupts. This technique works by
marking one of the registers with a special value, that the interrupt handling
code will recognise and roll the PC back so that the sequence is restarted from
the beginning, thus ensuring that the sequence is always executed atomically.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>