stm_fdma: Rewrite of FDMA driver
This is a major rework of the FDMA driver. This is an intermediate release
to get feedback on the revisised API, and provide support for the 7200.
Changes are partly structural:
 - better separation of platform initialisation data from that needed by
   the drivers
 - common up FDMA board support initialisation code
 - remove any visibility of FDMA specific data structures from the
   generic API

Also changes to the API:
 - remove all publicaly visible typedefs
 - try and employ consistent naming and spelling for all visable APIs
 - try to better indicate which functions are parameter manipulation
   functions and which are not

Changes to the request line handling so that now a driver must explicitly
request the use of a req line (which may fail if it is already in use)
and initialise its parameters.

Plus general code tidy up, optimistaion and debug.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>