sh_stm: Tidy up of BSPs and interrupt handling for the STMicro boards
Further cleanup of the BSP layer for the STMicro boards. Also
introduces a LED driver and changes from the old SA_INTERRUPT 
flag style to the new. The ASC and PIO drivers were SA_INTERRUPT,
this patch changes them to be normal interrupts, ie interrupts
are enabled in their interrupt handlers.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>