sh: Correct handling of misaligned faults in exception tables
When handling a misaligned access in the kernel, if the faulting instruction
has an entry in the exception table, the PC will already have been adjusted
to point to the fixup table. In this case we need to prevent the misaligned
fixup code adjusting the PC again on the assumption that it still points to
the faulting instruction which needs to be skipped.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>