alsa_stm: Fix left/right swap in STb7100 PCM driver
This patch inverts the clock edge bit determining LR PCM channel
ordering for 32bit mode.  In doing so we break LR ordering of 16b
playback due to our packing of a stereo pair of 16b samples into a
single I2s 32b word (necessary due to fixed external DAC config of
I2s). This is because packing into buffer is done LSB first, read out
is vice versa). The original situation left us with by luck left
channel packed and L sample on falling edge. Now we have left channel
packed first, but on rising edge, so the L channel is clocked out as
R.

To prevent this we switch back to FALLING edge clock for 16b mode in
the knowledge that we will lose the first stereo sample, meaning our
output stage is 1 sample out of phase from the position SW expects. I
am prepared to live with the guilt of this :)
 
An additional problem which was highlighted by the above fix was that,
when swtiching between 16/32 - 32/16 bit modes on PCM0 / PCM_CNV (off
chip DAC in both cases) there was an unreliable channel switch which
took place if the clocks were not removed after each playback.  Alas
removing the clocks is oft fatal to both the converter IP and Ext DAC
So we temporarily also disable 16b mode here for PCM0 & PCM_CNV until
a solution is found.

This patch addresses bugzilla bugs 1059 & 1066.

Signed-off-by: Mark Glaisher <mark.glaisher@st.com>