smsc911x: Add support for scatter-gather DMA
This patch modifies the generic SMSC 911x driver to allow it to make use
of DMA hardware which can support scatter/gather operation, in particular
the FDMA on the STMicroelectronics 710x devices. Currently this is only
used on packet reception.

This patch has a tendency to increase the number of packets which are
buffered in the SMSC's internal FIFO, and thus makes it more likely
to hit bugs present in the rev A chips. For this reason the driver
effectivly drops back to the previous behaviour on rev A chips.

It is assumed that the 911x's FIFOSEL pin is connected to A16.

A previous version of this patch was released under the name
smsc911x-fdma2.patch.

Note this patch needs fixing for multi-instance use.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>