mm_bpa2: Add BPA2, an improved version of the bigphysarea patch
This is a replacement for the bigphysarea patch. As well as the original
features (the ability to allocate a block of memory from the bootmem
allocator, and then make allocations from it at run time) it adds the
ability to handle multiple partitions, and have them at fixed
physical locations.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>