sh: Fix get_user(to,from++) usage
When a get_user(to, from++) is called the pointer increment is
performed after its first usage, in the specific after the __add_ok
invokation. This causes a wrong get_user return value, putting a wrong
character in the destination variable.This patch solves the problem
using a new temporary pointer.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>