sh: Export min/max_low_pfn to allow use of pfn_valid in modules
This patch allows pfn_valid from linux/asm-sh/page.h to be used in modules,
required by a driver's mmap nopage handler. It is an almost exact copy of 
the IA64 code, except the config option it is conditional on is the same as
that used to conditionally define pfn_valid on the SH.

It also exports fb_destroy_modelist for stmfb.

Signed-off-by: Andre Draszik <andre.draszik@st.com>
Signed-off-by: Stephen Gallimore <stephen.gallimore@st.com>