sh: Change uImage target to a wrapped compressed vmlinux
This patch changes the uImage target so that it generates a wrapped compressed
vmlinux, rather than a wrapped zImage. The previous version matched the ARM,
this version matches the PPC. However I would question how useful a self
decompressing image is with a boot loader which does decompression, so I
think this is more useful. I also feel it matches the descrition in the help
text ("Compressed kernel image") better.

Signed-off-by: Thomas Betker <thomas.betker@5etech.eu>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>