usb_stm: Workaround for bitstuffing problem on STmicro 710x chips
Checksums with a large number of 1's are occasionaly producing extra bits on
the checksum itself. It appears that SMSC hubs are the only ones that will
complain about this. Fortunately, there is a very simple workaround for this.
which is to set the number of retries on error to infinite (ie a 0 value for
the CERR field). Eventually the packet will succeed, and all is well.

Signed-off-by: David McKay <david.mckay@st.com>