sh: Fix problem with pud handling in do_page_fault()
This fixes a problem in the code which copies the vmalloc portion of the
kernel's page table into the current user space page table. The addition
of the four level page table code breaks on folded page tables, because
the pud level is always present (although folded). This updates the code
to use the same style of updates for the pud as is used for the pgd level.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>