sh: Change implementation of consistent_alloc() again
This is a second attempt to fix the problems seen with consistent_alloc()
when 32 bit mode is enabled. The first attempt was in
linux-sh4-2.6.23.13_stm23_0106-consistent-32bit-fixes.patch.

The problem is that consistent_alloc() can called from interrupt level
(with GFP_ATOMIC flags of course). So while the previous patch
allowed GFP_ATOMIC to be used, when called from interrupt level this hit
the in_interrupt() test in __get_vm_area_node().

So this patch uses the same technique as ARM and PowerPC, dedicating a
separate portion of virtual address space which is used only for
consistent mappings, and managed here.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>