sh: Workaround for SH4-300 delay slot double write bug
This patch works around a bug found in the SH4-300 (GNBvd67168). This causes
stores in delay slots to not be cancelled when interrupted. This makes
stores to FIFOs and other devices which are sensitive to how many times
the data is written unsafe in these circumstances.

This patch simply moves the store out of the delay slot. This is unlikely
to have any impact on performance on an SH4, although an SH3 would see one
additional cycle per loop iteration.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>