stm-rng: Hardware RNG /dev/random support:
Adds a driver which periodically fills the random number
entropy pool with values obtained from the hardware random number
generator present on the newer STM SoCs (7109 onwards)

(This allows /dev/random (which normally requires mouse and
keyboard connection to obtain entropy) to work on embedded systems and
provide cryptographically strong random numbers.  It can be also be
used in conjunction with the hw_random driver if required).

Signed-off-by: Carl Shaw <carl.shaw@st.com>