smc91x: Improve phy powerup code for SMSC LAN91C11 ethernet
According to the SMSC LAN91C11 datasheet, the device is guaranteed to be 
ready for normal operation 500ms after power-down is de-asserted.
The smc_drv_resume function invokes the smc_phy_configure but it seems 50ms are
not to sufficient to put the PHY in isolation mode.
Indeed, the driver requires to wait 50ms after the reset pin was de-asserted.
This patch reviews the resume function.
Instead of calling the smc_phy_configure now the driver follows the procedure 
described into the datasheet.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>