sh_stm: Rework STmicro USB chip specific configuration
This patch reworks the USB configuration for all the STM chips so that
instead of using #ifdefs based on the various chips, we pass in a flag
parameter to the wrapper configuration which describes the particular style
of configuration required for that chip.

This is primarily in preparation for the 7141, which has two different
styles of wrappers on the same chip.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>