serial_stasc: Don't wait for characters to drain when setting termios
Remove the loop which waiting for all tx characters to drain before
applying the new termios. If this functionality is needed it will have
been handled at a higher level, and if it wasn't requested, we shouldn't
do it. It also breaks in cases where flow control has been enabled but
CTS is not active.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>