phy: fix a failure reading zero from the UID PHY registers
Using a bad PHY address, the PHY device is erroneously attached to the MAC.
Unfortunately, broken hardware is sometimes missing the pull down resistor
on the MDIO line, which results in reads to non-existent devices returning
0 rather than 0xffff from the UID PHY registers (see the get_phy_id function).
This is caught as soon as the PHY device is attached to the MAC where phy_id=0
is treated as a non-existent device.

See STLinux Bugzilla Bug #4967 too.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>