fix section mismatch warnings compiling the nwhwconfig driver.
If the NETPOLL support is enable, we get the following message building the
Kernel image:

WARNING: vmlinux.o(.text+0x19b050): Section mismatch: reference to
.init.text:set_reset_devices (between 'nwhw_uconfig' and 'scsi_device_type')

It is due to the nwhw_uconfig definition.
So this patch just fixes that.
Note: nwhw_uconfig function was added to provide the MAC address while
forcing the device open by using he netpoll_setup function.
It is required when kgdboe uses the stmmac, for example.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>