stm_fdma: Fix race condition in stb710x_fdma_get_residue
Updated patch following suggestion from Pawel Moll that we write
the length directly into the hardware register, rather than use an
intermediate value.

A problem was seen when running multiple FDMA channels concurrently.
In this case it appears to be possible to write to the registers to
start a DMA transfer, return and call get_residue() before the FDMA
has even started the transfer. In this case the code reported that the
transfer had completed, because it was reading the values left over
from the previous transfer.

So be much more careful about possible race conditions between the FDMA
and the CPU:
 - when reading the FDMA registers loop to ensure we get a consistent
   status and count (status also contains the node pointer)
 - check the channel status, and if we've gone idle but haven't yet taken
   the interrupt return a length of 1.
 - write the full length into the count register before starting the
   transfer, so that if the hardware hasn't updated it yet we still get
   the correct value.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Acked-by: Pawel Moll <pawel.moll@st.com>