i2c-stm: I2C no repeated start support (with STOP START delay)
This patch adds support to the I2C driver which forces messages
within a transaction to be separated by a STOP then START rather
than be a REPEATED START.

In a multi-master system, this will allow another master to get
control of a bus in the middle of a long transaction.

Where the host is the only master, this patch can fix issues
with some I2C tuner chips which want all transactions separated
by a STOP in order to make the I2C repeater work.

Signed-off-by: Carl Shaw <carl.shaw@st.com>
Reviewd-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>