stmmac: update the driver to the latest version
This patch updates the stmmac driver to the latest version
will be released into the 118 Kernel tree.

It requires the previous patch sent in September named:
"Add the stmmac network device driver" (always for the wavefront tree).

It adds the experimental support to perform emulated hardware segmentation
on skb. It can be enabled by using the ethtool command and is not fully
tested.

It fixes a known bug within the MDIO reset as well.

Finally, it replaces the simple_strtoul with strict_strtoul within the
stmmac_cmdline_opt function.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>