sh_stm: Add support for the mb680 revision C
The mb680 rev C switched the Ethernet PHY from a NatSemi DP83865
to an SMSC 8700. It also added software control of the PHY reset, and
software controlled switches to isolate the PHY from the 7105, to
overcome problems with both devices using the same lines for mode
selection.

See bugzilla 5441 and 5482.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>