stmmac: use first detected phy
This change allows the stmmac's platform data to specify the PHY address
as -1, in which case the first available PHY on the bus will be used.

This avoids problems when the PHY address changes between board revisions
for example. Also output some diagnostic messages which can be useful to
help diagnose PHY problems.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>