serial_stasc: handle stopped ports better
The previous code only checked whether the port had been stopped if
there was room in the Tx buffer. This could lead to a race condition
if the port had been closed and thus was stopped. In this case if
there was no space in the Tx FIFO the code would still try and call
uart_write_wakeup, causing a NULL Pointer dereference.

See bugzilla 5253.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>