stm_fdma: Fix stb710x_fdma_get_residue for circular lists
The previous modification to stb710x_fdma_get_residue (commit
916f97b0e408fdb50ed58beb9daa3b2d347f1daf "stm_fdma: Fix reading of
addresses from hardware in stb710x_fdma_get_residue") introduced a bug
in the case of circular lists. This patch fixes that bug, and should
also be more efficient, as it removes the read of the node's next
pointer, which is uncached.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>