sound_stm: Ignore interrupts when the player/reader is inactive
For some applications we have a seperate sound driver operating on
another processor. In these applications the system 'context switches'
between the host sound drivers and the remote ones. We must therefore
park the host drivers such that they are not affected if the players
and readers generate interrupts when under the control of a different
processor.

Signed-off-by: Daniel Thompson <daniel.thompson@st.com>
Signed-off-by: Pawel Moll <pawel.moll@st.com>