rt73_wireless: boottime firmware loading
A new firwmare loading strategy is implemented allowing a user to choose
to include RT73 firmware directly with the kernel .text section so that it
is available at boot.

Enable the kernel configuration option "Rt73 firmware loaded inside kernel"
and put the rt73.bin firmware under the usr directory of the kernel tree.

With this patch it is now possible to boot a kernel with a wireless device
plugged on usb.  Upon USB unloading and reloading, the standard firmware
(in /lib/firmware) will be used.

Signed-off-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>