libata: Keep shadow last_ctl up to date during resets
libata keeps a shadow copy of the ATA CTL register (which is write only),
and only writes to the hardware when the required value doesn't match
the shadow. However this copy wasn't being maintained when performing
reset functions. This could cause problems for the first operation after
a reset when the correct value might not be written to the CTL register.

See bugzilla https://bugzilla.stlinux.com/show_bug.cgi?id=5391

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>