sh_stm: Add support for ST40-300 as an explicit sub-architecture
Now that gcc supports ST40-300 as an explicit architecture variant,
add support into the kernel build. Currently this simply selects the
compilation option when building, which may allow for slightly more
optimal code generation. However it also makes CONFIG_CPU_ST40_300
available as an config option which can be used in the future.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>