sh_stm: Fix a race condition updating page flags
Fix a race condition in the code which handles cache flushing
to work around D-cache aliasing between kernel and user space. The
page flag which marks the page as dirty, and in need of flushing,
wasn't being manipulated using the atomic set and clear macros, and so
sometimes raced with other functions manipulating the page flags.

See bugzilla 5885

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>