sh: Fix hard IRQ tracing
There are a few missing calls in the SH assembler code to keep the hard
IRQ tracing code informed about that state of the hardware's IRQ mask.

Also switch to using a macro which removes quite a few #ifdefs. This
is slightly less efficient because we have to jump over the function
address, but given that its only used when tracing is enabled its
probably acceptable.

See bugzilla 6138

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>