stm_pio: Fix STMicro gpio_direction_output()
The generic gpio_direction_output() takes a "value" parameter which is
the state you want to set the pin to. The current STMicro STPIO
implementation completely ignores this and doesn't set any value at all,
so you end up with whatever it was. The patch fixes this by setting the
value before you configure the pin as an output, thus avoiding any
glitches.

Signed-off-by: David McKay <david.mckay@st.com>