serial_stasc: Fix lockup when Rx buffer full
Prior to this patch, a lockup could occur when the software Rx buffer
became full. In this case any further received data would cause an
interrupt, but the data wouldn't have been read from the FIFO, leaving
the interrupt asserted.

Removing the check for the buffer being full means that the FIFO is
now read in all cases, and the data discarded by tty_insert_flip_char()
if there is insufficient room. This now matches the behaviour of most
other serial drivers.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>