sh_stm: Support PCI reset via PIO pin
Many STMicro boards use a PIO pin to control the PCI reset line. This
patch adds support for this case, and also allows the user to provide a
board specific PCI reset function in case the PCI reset is done via some
other mechanism. The mb680 7105 board actually uses a PIO line, so this
patch may fix some problems with some cards as the reset was never done
on this board previously.

Signed-off-by: David McKay  <david.mckay@st.com>