sh_stm: Don't write to sysconfig for PIO pins which don't need it
Some PIO pins do not require sysconfig level configuration to select
which alternative function will drive the pin. This has always been
flagged in the pio_sysconf array with a syscfg field of 0. However this
caused any configuration of these pins to write to sysconf config 0,
which was incorrect. So add an explicit check for this in
stx7105_pio_sysconf() and return immediately in this case.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>